+++ test/libsolidity/gasTests/abiv2.sol
@@ -18,7 +18,7 @@ contract C {
 //   executionCost: 1160
 //   totalCost: 1121160
 // external:
-//   a(): 530
+//   a(): 1130
 //   b(uint256): infinite
 //   f1(uint256): infinite
 //   f2(uint256[],string[],uint16,address): infinite